pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ww"
version = "0.1.0"
description = "A cross-platform CLI toolkit for personal and enterprise automation, with LLM integration."
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "requests>=2.32.0",
    "pyperclip>=1.8.0",
    "markdown>=3.5.0",
    "beautifulsoup4>=4.12.0",
    "readability-lxml>=0.8.1",
    "lxml>=5.0.0",
    "Pillow>=10.0.0",
    "numpy>=1.24.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
ww = "ww.main:main"

[tool.pyright]
pythonVersion = "3.8"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingModuleSource = false

[tool.ruff.lint]
ignore = [
    "E731",  # lambda assignments — used intentionally as prompt builders
    "E722",  # bare except — intentional in network/system code
    "F841",  # unused local variable
]

[tool.ruff.lint.per-file-ignores]
"ww/search/*.py" = ["E402"]  # warnings.filterwarnings must precede imports
